requirements.txt
numpy
scipy
pyapriltags
open3d
opencv-python
pillow
pyrealsense2